requirements.txt
beautifulsoup4==4.12.3
google_api_python_client==2.117.0
httplib2==0.22.0
oauth2client==4.1.3
protobuf==4.25.3
python-dotenv==1.0.1
Requests==2.31.0
